Moved `isStoredInSecretStorage` into a single function
